# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: GlobalBooks OrdersService API
  version: "1.0.0"
  description: |
    REST API for order management in the GlobalBooks SOA.
    Secured with OAuth2 JWT Bearer tokens (issued by Keycloak).
    Q7: REST API specification
  contact:
    name: SOA Team
    email: soa-team@globalbooks.com

servers:
  - url: http://localhost:8081/api/v1
    description: Local development

security:
  - bearerAuth: []

paths:
  /orders:
    post:
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customerId: "C1001"
              items:
                - bookId: "B001"
                  quantity: 2
                - bookId: "B003"
                  quantity: 1
              shippingAddress:
                street: "1 Book Lane"
                city: "London"
                country: "UK"
                postalCode: "EC1A 1BB"
      responses:
        '201':
          description: Order created successfully
          headers:
            Location:
              schema:
                type: string
              description: URI of the created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized – missing or invalid JWT token

    get:
      summary: List all orders
      operationId: getAllOrders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized

  /orders/{id}:
    get:
      summary: Get an order by ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "ORD-20260323-001"
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
        '401':
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token from Keycloak – POST to /realms/globalbooks/protocol/openid-connect/token"

  schemas:
    CreateOrderRequest:
      type: object
      required: [customerId, items]
      properties:
        customerId:
          type: string
          example: "C1001"
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'

    OrderItem:
      type: object
      required: [bookId, quantity]
      properties:
        bookId:
          type: string
          example: "B001"
        title:
          type: string
          readOnly: true
          example: "Clean Code"
        quantity:
          type: integer
          minimum: 1
          maximum: 99
          example: 2
        unitPrice:
          type: number
          format: decimal
          readOnly: true
          example: 31.99

    ShippingAddress:
      type: object
      required: [street, city, country, postalCode]
      properties:
        street:
          type: string
          example: "1 Book Lane"
        city:
          type: string
          example: "London"
        country:
          type: string
          example: "UK"
        postalCode:
          type: string
          example: "EC1A 1BB"

    Order:
      type: object
      properties:
        orderId:
          type: string
          example: "ORD-20260323-001"
        customerId:
          type: string
          example: "C1001"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: decimal
          example: 95.97
        currency:
          type: string
          example: "USD"
        status:
          type: string
          enum: [PENDING, PAID, SHIPPED, DELIVERED, CANCELLED]
          example: "PENDING"
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        createdAt:
          type: string
          format: date-time
          example: "2026-03-23T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-23T10:00:00Z"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: string
          example: "items must contain at least one entry"